fix: fix metadata format and sequence processing (#15)
This pull request makes two fixes: it updates the output metadata file
format to TSV and fixes a bug in the lineage processing logic to ensure
as many samples are processed.

**Output format update:**

* Changed the default output metadata file from CSV to TSV and updated
the help text in the `barcode` CLI command to reflect this change.

**Bug fix in lineage processing:**

* Replaced a `break` with `continue` in the
`process_and_reroot_lineages` function so that all missing samples are
skipped individually, rather than stopping at the first missing sample.